requirements.txt
Flask==0.10.1
Frozen-Flask==0.11
Jinja2==2.7.1
MarkupSafe==0.18
Werkzeug==0.9.3
boto==2.10.0
itsdangerous==0.23
requests==1.2.3
wsgiref==0.1.2
Pillow==2.8.1
